fix for intToString

+++ REAL/TPE-Arqui/Userland/SampleCodeModule/include/stdlib.h
@@ -1,8 +1,11 @@
 #ifndef STDLIB_H
 #define STDLIB_H
 
+#include <syscallsASM.h>
+#include <stddef.h>
+
 int atoi(char * str);
-int intToString(unsigned long long num, char * buffer);
+int intToString(uint64_t num, char * buffer);
 int strlen(char *str);
 int strcmp(char * str1, char * str2);
 int intToBase(unsigned long long num, int base, char*buffer);
